+++ app/contact/head.tsx
@@ -1,0 +1,25 @@
+export default function Head() {
+  return (
+    <>
+      <title>Contact - Nate Reprogle</title>
+      <link rel="canonical" href="https://reprogle.org/contact" />
+      <meta property="og:locale" content="en_US" />
+      <meta property="og:type" content="article" />
+      <meta property="og:title" content="Contact - Nate Reprogle" />
+      <meta
+        property="og:description"
+        content="Send me a message. Whether you wish to just chat or need to reach me, simply fill this form and I’ll get back to you soon."
+      />
+      <meta property="og:url" content="https://reprogle.org/contact" />
+      <meta property="og:site_name" content="Nate Reprogle" />
+      <meta
+        property="article:modified_time"
+        content="2022-10-29T15:18:00+00:00"
+      />
+      <meta property="og:image" content="https://reprogle.org/Contact.jpg" />
+      <meta name="twitter:card" content="summary_large_image" />
+      <meta name="twitter:label1" content="Est. reading time" />
+      <meta name="twitter:data1" content="1 minute" />
+    </>
+  );
+}
