+++ app/keys/head.tsx
@@ -1,0 +1,35 @@
+export default function Head() {
+  <>
+    <title>Keys - Nate Reprogle</title>
+    <meta
+      name="description"
+      content="Public SSH & PGP Keys used by Nate Reprogle"
+    />
+    <link rel="canonical" href="https://reprogle.org/keys" />
+
+    <meta property="og:locale" content="en_US" />
+    <meta property="og:type" content="website" />
+    <meta property="og:title" content="Home - Nate Reprogle" />
+    <meta
+      property="og:description"
+      content="Public SSH & PGP Keys used by Nate Reprogle"
+    />
+    <meta property="og:url" content="https://reprogle.org/contact" />
+    <meta property="og:site_name" content="Nate Reprogle" />
+    <meta
+      property="article:modified_time"
+      content="2022-10-29T15:18:00+00:00"
+    />
+    <meta property="og:image" content="https://reprogle.org/Profile.jpg" />
+
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:site" content="@NateReprogle" />
+    <meta name="twitter:creator" content="@NateReprogle" />
+    <meta name="twitter:title" content="Home - Nate Reprogle" />
+    <meta
+      name="twitter:description"
+      content="Public SSH & PGP Keys used by Nate Reprogle"
+    />
+    <meta name="twitter:image" content="https://reprogle.org/Profile.jpg" />
+  </>;
+}
