+++ components/contactForm.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { useState } from "react";
 import { Turnstile } from "@marsidev/react-turnstile";
