+++ components/keys.tsx
@@ -1,6 +1,7 @@
+'use client';
 import { useState } from "react";
 
-function Keys({ ...props }) {
+export default function Keys({ ...props }) {
   const expandArrow = (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -197,5 +198,3 @@ function Keys({ ...props }) {
     </div>
   );
 }
-
-export default Keys;
