--- pages/_document.tsx
@@ -1,13 +1,0 @@
-import { Html, Head, Main, NextScript } from "next/document";
-
-export default function Document() {
-  return (
-    <Html>
-      <Head />
-      <body className={"bg-dark-base-blue-100"}>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-}
